keep hyperref links when adding page numbers fixes Issue #2
This commit uses a new method to add pages numbers and citation stamps
to the author-provided pdfs in order to prevent the clickable links in
those files from being broken.

The user must install the python library pdfrw from:
https://github.com/pmaupin/pdfrw

This library is used by the new script pdfunderneath.py,
which merges two pdfs page by page.  It is called by manage-db.pl,
which calls pdflatex to create a pdf containing only page numbers
and the citation stamp, and then merges this file with the
author-provided pdf using pdfunderneath.py.

The links still do not work in the file "book.pdf" that contains all
of the papers.  This file contains its own clickable links in the
table of contents and index, and I have not found any way of getting
both sets of links working at the same time.